refactor(cli): extract source render helpers to _source_render.py (#1331)
Move the render/validation helper bank (the 17 helpers `_looks_like_path`
… `_render_source_add_drive_result`) and the four add-research render
helpers (`_emit_add_research_flag_conflict`, `_print_add_research_task_ids`,
`_exit_with_add_research_status`, `_render_add_research_result`) out of
`cli/source_cmd.py` into a new `cli/_source_render.py`. Bodies are moved
verbatim; every helper is re-exported from `source_cmd` so the historical
`source_cmd.<helper>` import/patch surface keeps resolving.

No command body moves — all 137 `source_cmd.NotebookLMClient` patch seams
stay put. The three dependency-rebind tests in
`tests/unit/test_cli_source_services.py` that `monkeypatch.setattr` the
moved `json_output_response`/`display_research_sources`/`display_report`
names are repointed to `_source_render` (the `source_cmd.console.print`
patch survives — `console` is a shared object). The 8 direct-call
add-research tests in `test_source_cmd_coverage.py` survive via re-export.

source_cmd.py: 1498 → 947 LOC (ratchet ceiling lowered 1498 → 947, stays
allowlisted). New `_source_render.py`: 609 LOC (≤ 900, no entry needed).
No public-API change; audit_public_api_compat exits 0 with 0 new
allowlist entries.
