+++ tests/unit/test_cli_source_services.py
@@ -9,7 +9,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from notebooklm.cli import source_cmd
+from notebooklm.cli import _source_render, source_cmd
 from notebooklm.cli.services import source_mutations, source_research
 from notebooklm.cli.services.source_content import (
     SourceFulltextPlan,
@@ -331,9 +331,9 @@ def test_render_add_research_started_no_wait_json_payload(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     payloads: list[dict[str, object]] = []
-    monkeypatch.setattr(source_cmd, "json_output_response", payloads.append)
+    monkeypatch.setattr(_source_render, "json_output_response", payloads.append)
 
-    source_cmd._render_add_research_result(
+    _source_render._render_add_research_result(
         SourceAddResearchResult(
             outcome="started_no_wait",
             plan=SourceAddResearchPlan(
@@ -366,14 +366,14 @@ def test_render_add_research_completed_json_payload(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     payloads: list[dict[str, object]] = []
-    monkeypatch.setattr(source_cmd, "json_output_response", payloads.append)
+    monkeypatch.setattr(_source_render, "json_output_response", payloads.append)
     import_result = SimpleNamespace(
         imported=[{"id": "src_1", "title": "Result"}],
         cited_selection=SimpleNamespace(used_fallback=False),
         sources=[{"title": "Result"}],
     )
 
-    source_cmd._render_add_research_result(
+    _source_render._render_add_research_result(
         SourceAddResearchResult(
             outcome="completed",
             plan=SourceAddResearchPlan(
@@ -419,10 +419,10 @@ def test_render_add_research_completed_text_keeps_task_ids(
     monkeypatch.setattr(
         source_cmd.console, "print", lambda message="", *_, **__: printed.append(message)
     )
-    monkeypatch.setattr(source_cmd, "display_research_sources", lambda sources: None)
-    monkeypatch.setattr(source_cmd, "display_report", lambda report, json_hint=False: None)
+    monkeypatch.setattr(_source_render, "display_research_sources", lambda sources: None)
+    monkeypatch.setattr(_source_render, "display_report", lambda report, json_hint=False: None)
 
-    source_cmd._render_add_research_result(
+    _source_render._render_add_research_result(
         SourceAddResearchResult(
             outcome="completed",
             plan=SourceAddResearchPlan(
